Crate teaclave_attestation

This crate provides TLS-based remote attestation mechanism for Teaclave, supporting both EPID and ECDSA attestation. By default, Intel Attestation Service is used for RA.

Modules

report

Types that contain information about attestation report. The implementation is based on Attestation Service API version 4. https://api.trustedservices.intel.com/documents/sgx-attestation-api-spec.pdf

verifier

This module provides types used to verify attestation reports.

Structs

AttestationServiceConfig

Attestation Service Configuration

AttestedTlsConfig

Configuration for TLS communication in Remote Attestation

DcapConfig

EndorsedAttestationReport

AttestationReport can be endorsed by either the Intel Attestation Service using EPID or Data Center Attestation Service (platform dependent) using ECDSA.

Enums

AttestationConfig

Remote attestation configuration

AttestationError

Errors that can happen during attestation and verification process