Module teaclave_attestation::report

Types that contain information about attestation report. The implementation is based on Attestation Service API version 4. https://api.trustedservices.intel.com/documents/sgx-attestation-api-spec.pdf

Structs

AttestationReport

A report that can be signed by Intel EPID (which generates EndorsedAttestationReport) and then sent off of the platform to be verified by remote client.

SgxEnclaveReport

A report generated by an enclave that contains measurement, identity and other data related to enclave.

SgxQuote

An application that hosts an enclave can ask the enclave to produce a report (SgxEnclaveReport) and then pass this report to a platform service (Quoting Enclave) to produce a type of credential that reflects the enclave and platform state. The quote can be passed to entities off the platform, and verified using Intel EPID signature verification techniques.

Enums

SgxEcdsaQuoteAkType

ECDSA attestation key type

SgxEpidQuoteSigType

Intel EPID attestation signature type

SgxQuoteStatus

SGX Quote status

SgxQuoteVersion

SGX Quote structure version