Struct teaclave_attestation::report::SgxQuote

pub struct SgxQuote {
    pub version: SgxQuoteVersion,
    pub gid: u32,
    pub isv_svn_qe: u16,
    pub isv_svn_pce: u16,
    pub qe_vendor_id: Uuid,
    pub user_data: [u8; 20],
    pub isv_enclave_report: SgxEnclaveReport,
}

An application that hosts an enclave can ask the enclave to produce a report (SgxEnclaveReport) and then pass this report to a platform service (Quoting Enclave) to produce a type of credential that reflects the enclave and platform state. The quote can be passed to entities off the platform, and verified using Intel EPID signature verification techniques.

Fields
version: SgxQuoteVersion

Version of the quote structure

gid: u32

ID of the Intel EPID group of the platform belongs to

isv_svn_qe: u16

Security version number of Quoting Enclave

isv_svn_pce: u16

Security version number of PCE

qe_vendor_id: Uuid

Vendor ID of Quoting Enclave

user_data: [u8; 20]

User data

isv_enclave_report: SgxEnclaveReport

Report generated by the enclave

Implementations

impl SgxQuote

pub fn parse_from<'a>(bytes: &'a [u8]) -> Result<Self>

Parse from bytes to SgxQuote.