Struct teaclave_attestation::verifier::AttestationReportVerifier

pub struct AttestationReportVerifier {
    pub accepted_enclave_attrs: Vec<EnclaveAttr>,
    pub root_ca: Vec<u8>,
    pub verifier: AttestationReportVerificationFn,
}

Type used to verify attestation reports (this can be set as a certificate verifier in rustls::ClientConfig).

Fields
accepted_enclave_attrs: Vec<EnclaveAttr>

Valid enclave attributes (only enclaves with attributes in this vector will be accepted).

root_ca: Vec<u8>

Root certificate of the attestation service provider (e.g., IAS).

verifier: AttestationReportVerificationFn

User defined function to verify the attestation report.

Implementations

impl AttestationReportVerifier

pub fn new(
accepted_enclave_attrs: Vec<EnclaveAttr>,
root_ca: &[u8],
verifier: AttestationReportVerificationFn
) -> Self